@@ -110,19 +110,7 @@ module.exports = {
             add_committee_label: "理事会成员",
             add_witness_label: "见证人",
             approve_worker: "赞成",
-            reject_worker: "反对",
-            worker_account: "申请账户",
-            total_votes: "赞成票数",
-            votes_against: "反对票数",
-            daily_pay: "每日预算",
-            max_pay: "最大预算",
-            unclaimed: "未发放预算",
-            status: {
-                title: "我的立场",
-                supported: "已支持",
-                rejected: "已反对",
-                neutral: "中立"
-            }
+            reject_worker: "反对"
         },
         options: {
             num_witnesses: "信任见证人",
@@ -154,7 +142,13 @@ module.exports = {
         identicon: "身份标示图形",
         pay_from: "注册费支付账户",
         existing_accounts: "已有账户",
-        bts_market: "交易市场"
+        bts_market: "交易市场",
+        name_input: {
+            name_is_taken: "Account name is already taken.",
+            not_found: "Account not found.",
+            premium_name_faucet: "This is a premium name. Premium names are more expensive and can't be registered for free by faucet. Try to select another name containing at least one dash, number or no vowels.",
+            premium_name_warning: "This is a premium name that is more expensive to register. Regular names have at least one dash, number or no vowels."
+        }
     },
     transfer: {
         from: "来自",
@@ -289,17 +283,7 @@ module.exports = {
             balance_claim: "领取余额",
             override_transfer: "优先覆盖转账",
             witness_update: "更新见证人",
-            committee_member_update_global_parameters: "全局参数更新",
-            vesting_balance_create: "创建冻结余额",
-            vesting_balance_withdraw: "提取解冻余额",
-            worker_create: "创建预算项目",
-            custom: "自定义",
-            assert: "断言操作",
-            balance_claim: "领取余额",
-            override_transfer: "覆盖转账",
-            transfer_to_blind: "向隐私账户转账",
-            blind_transfer: "隐私转账",
-            transfer_from_blind: "从隐私账户转出"
+            committee_member_update_global_parameters: "全局参数更新"
         },
         confirm: "请确认交易",
         broadcast_success: "交易已向网络广播",
@@ -321,12 +305,7 @@ module.exports = {
         new_url: "网站",
         publisher: "发布人",
         lifetime_upgrade_account: "升级到终身会员",
-        annual_upgrade_account: "升级到年度会员",
-        market_fee: "市场手续费",
-        max_market_fee: "最大市场手续费",
-        blinding_factor: "隐私因子",
-        outputs: "输出",
-        inputs: "输入",
+        annual_upgrade_account: "升级到年度会员"
     },
     explorer: {
         accounts: {
@@ -370,9 +349,9 @@ module.exports = {
             listed_account: "Listed account",
             new_listing: "New listing",
             asset_update: "可更新资产",
-            common_options: "参数设定",
-            new_options: "新的设定",
-            new_producers: "新的喂价提供者",
+            common_options: "Common options",
+            new_options: "New options",
+            new_producers: "New feed producers",
             asset_issue: "发行数量",
             max_margin_period_sec: "Max margin period (s)",
             call_limit: "Call limit",
