@@ -140,7 +140,13 @@ module.exports = {
         create_account: "Konto erstellen",
         identicon: "Identicon",
         pay_from: "Bezahlen von",
-        existing_accounts: "Existierendes Konto"
+        existing_accounts: "Existierendes Konto",
+        name_input: {
+            name_is_taken: "Account name is already taken.",
+            not_found: "Account not found.",
+            premium_name_faucet: "This is a premium name. Premium names are more expensive and can't be registered for free by faucet. Try to select another name containing at least one dash, number or no vowels.",
+            premium_name_warning: "This is a premium name that is more expensive to register. Regular names have at least one dash, number or no vowels."
+        }
     },
     pagination: {
         newer: "Jüngere",
