@@ -141,7 +141,13 @@ module.exports = {
         create_account: "계정 생성",
         identicon: "아이덴티콘",
         pay_from: "Pay From",
-        existing_accounts: "기존 계정"
+        existing_accounts: "기존 계정",
+        name_input: {
+            name_is_taken: "Account name is already taken.",
+            not_found: "Account not found.",
+            premium_name_faucet: "This is a premium name. Premium names are more expensive and can't be registered for free by faucet. Try to select another name containing at least one dash, number or no vowels.",
+            premium_name_warning: "This is a premium name that is more expensive to register. Regular names have at least one dash, number or no vowels."
+        }
     },
     pagination: {
         newer: "이전",
