@@ -2,6 +2,7 @@ import React from "react";
 import AccountSelector from "./AccountSelector";
 import BindToChainState from "../Utility/BindToChainState";
 import ChainTypes from "../Utility/ChainTypes";
+import Translate from "react-translate-component";
 
 @BindToChainState()
 class AccountVotingProxy extends React.Component {
@@ -15,7 +16,7 @@ class AccountVotingProxy extends React.Component {
     constructor(props){
         super(props);
         this.state = {
-            current_proxy_input: null,
+            current_proxy_input: props.currentProxy,
             new_proxy_account: null
         }
         this.onProxyChange = this.onProxyChange.bind(this);
@@ -26,6 +27,12 @@ class AccountVotingProxy extends React.Component {
         if(!this.state.current_proxy_input) this.setState({
             current_proxy_input: nextProps.proxyAccount ? nextProps.proxyAccount.get("name") : ""
         });
+
+        if (nextProps.currentProxy !== this.state.current_proxy_input) {
+            this.setState({
+                current_proxy_input: nextProps.currentProxy
+            });
+        }
     }
 
     onProxyChange(current_proxy_input) {
@@ -36,9 +43,9 @@ class AccountVotingProxy extends React.Component {
     onProxyAccountChange(new_proxy_account) {
         //console.log("-- AccountVotingProxy.onProxyAccountChange -->", new_proxy_account);
         if(new_proxy_account && this.props.currentAccount.get("id") === new_proxy_account.get("id"))
-            this.props.onProxyAccountChanged(null);
+            this.props.onProxyAccountChanged(null, this.state.current_proxy_input);
         else
-            this.props.onProxyAccountChanged(new_proxy_account);
+            this.props.onProxyAccountChanged(new_proxy_account, this.state.current_proxy_input);
         this.setState({new_proxy_account});
     }
 
@@ -48,8 +55,8 @@ class AccountVotingProxy extends React.Component {
         if(this.state.new_proxy_account && this.props.currentAccount.get("id") === this.state.new_proxy_account.get("id"))
             error = "cannot proxy to yourself";
         return (
-        <div className="content-block">
-            <h3>Proxy</h3>
+        <div className="content-block" style={{maxWidth: "600px"}}>
+            <h3><Translate content="account.votes.proxy_short" /></h3>
             <AccountSelector label="account.votes.proxy"
                              error={error}
                              account={this.state.current_proxy_input}
