@@ -6,6 +6,7 @@ import Immutable from "immutable";
 import AccountImage from "./AccountImage";
 import ChainStore from "api/ChainStore";
 import ChainTypes from "../Utility/ChainTypes";
+import FormattedAsset from "../Utility/FormattedAsset";
 import BindToChainState from "../Utility/BindToChainState";
 
 class AccountItemRow extends React.Component {
@@ -13,20 +14,43 @@ class AccountItemRow extends React.Component {
         account: React.PropTypes.object.isRequired,
         onRemoveItem: React.PropTypes.func.isRequired
     }
+
+    shouldComponentUpdate(nextProps) {
+        return nextProps.account !== this.props.account;
+    }
+
     onRemoveItem(item_id){
         this.props.onRemoveItem(item_id);
     }
+
     render() {
-        let name = this.props.account.get("name");
-        let item_id = this.props.account.get("id");
+        let {account, type} = this.props;
+        let name = account.get("name");
+        let item_id = account.get("id");
+
         let url = "";
+        let votes = 0;
+        if (type) {
+            if (type === "witness") {
+                let witness = ChainStore.getWitnessById(account.get("id"));
+                url = witness ? witness.get("url") : url;
+                votes = witness ? witness.get("total_votes") : votes;
+            } else if (type === "committee") {
+                let committee = ChainStore.getCommitteeMemberById(account.get("id"));
+                url = committee ? committee.get("url") : url;
+                votes = committee ? committee.get("total_votes") : votes;
+            }
+        }
+
+
         return (
-            <tr key={name}>
+            <tr>
                 <td>
                     <AccountImage size={{height: 30, width: 30}} account={name}/>
                 </td>
                 <td>{name}</td>
                 <td>{url}</td>
+                <td><FormattedAsset amount={votes} asset="1.3.0" decimalOffset={5} /></td>
                 <td>
                     <button className="button outline" onClick={this.onRemoveItem.bind(this, item_id)}>
                         <Translate content="account.votes.remove_witness"/></button>
@@ -98,35 +122,39 @@ class AccountsList extends React.Component {
                  return 0;
                  })
            .map(i => {
-            return (<AccountItemRow account={i} onRemoveItem={this.props.onRemoveItem}/>)
+            return (<AccountItemRow key={i.get("name")} account={i} type={this.props.type} onRemoveItem={this.props.onRemoveItem}/>)
            });
 
         let error = this.state.error;
         if(!error && this.state.selected_item && this.props.items.indexOf(this.state.selected_item) !== -1)
             error = "Account is already in the list";
 
-        let cw = ["10%", "20%", "60%", "10%"];
+        let cw = ["10%", "20%", "40%", "20%", "10%"];
 
         return (
             <div>
-                <AccountSelector label={this.props.label}
-                                 error={error}
-                                 placeholder={this.props.placeholder}
-                                 account={this.state.item_name_input}
-                                 accountName={this.state.item_name_input}
-                                 onChange={this.onItemChange}
-                                 onAccountChanged={this.onItemAccountChange}
-                                 onAction={this.onAddItem}
-                                 action_label="account.votes.add_witness"
-                                 tabIndex={this.props.tabIndex}/>
+                <AccountSelector
+                    style={{maxWidth: "600px"}}
+                    label={this.props.label}
+                    error={error}
+                    placeholder={this.props.placeholder}
+                    account={this.state.item_name_input}
+                    accountName={this.state.item_name_input}
+                    onChange={this.onItemChange}
+                    onAccountChanged={this.onItemAccountChange}
+                    onAction={this.onAddItem}
+                    action_label="account.votes.add_witness"
+                    tabIndex={this.props.tabIndex}
+                 />
                 <table className="table">
                     <thead>
-                    <tr>
-                        <th style={{width: cw[0]}}></th>
-                        <th style={{width: cw[1]}}><Translate content="account.votes.name"/></th>
-                        <th style={{width: cw[2]}}><Translate content="account.votes.url"/></th>
-                        <th style={{width: cw[3]}}>ACTION</th>
-                    </tr>
+                        <tr>
+                            <th style={{width: cw[0]}}></th>
+                            <th style={{width: cw[1]}}><Translate content="account.votes.name"/></th>
+                            <th style={{width: cw[2]}}><Translate content="account.votes.url"/></th>
+                            <th style={{width: cw[3]}}><Translate content="account.votes.votes" /></th>
+                            <th style={{width: cw[4]}}>ACTION</th>
+                        </tr>
                     </thead>
                     <tbody>
                     {item_rows}
