@@ -60,47 +60,51 @@ class WorkerApproval extends React.Component{
       } else if( !this.props.vote_ids.has( worker.vote_for ) && this.props.vote_ids.has( worker.vote_against ) ) {
          approval = counterpart.translate("account.votes.status.rejected");
       }
-      // console.log( "worker: ", worker );
 
       return  (
-      <div style={{padding: "0.5em 0.5em"}} className="grid-content account-card">
+      <div style={{padding: "0.5em 0.5em"}} className="grid-content account-card worker-card">
          <div className="card">
             <div className="card-divider text-center info">
                <span> {worker.name} </span>
             </div>
             <div className="card-section">
-            <ul >
-               <li>
-                  <span><Translate content="account.votes.worker_account" />:&nbsp;<LinkToAccountById account={worker.worker_account} /> </span>
-               </li>
-               <li>
-                  <div><Translate content="account.votes.url" />:&nbsp;{worker.url} </div>
-               </li>
-               <li>
-                  <Translate content="account.votes.total_votes" />: <FormattedAsset amount={total_votes} asset="1.3.0" /><br/>
-               </li>
-               <li>
-                  <Translate content="account.votes.daily_pay" />: <FormattedAsset amount={worker.daily_pay} asset="1.3.0" /><br/>
-               </li>
-               <li>
-                  <Translate content="account.votes.max_pay" />: <FormattedAsset amount={worker.daily_pay*total_days} asset="1.3.0" /><br/>
-               </li>
-               <li>
-                  <Translate content="account.votes.unclaimed" />: <VestingBalance balance={worker.worker[1].balance} /> <br/>
-               </li>
-               <li>
-                  <Translate content="account.votes.status.title" />: {approval} <br/>
-               </li>
-            </ul>
+               <ul >
+                  <li>
+                     <span><Translate content="account.votes.worker_account" />:&nbsp;<LinkToAccountById account={worker.worker_account} /> </span>
+                  </li>
+                  <li>
+                     <div><Translate content="account.votes.url" />:&nbsp;<a target="_blank" href={worker.url}>{worker.url}</a> </div>
+                  </li>
+                  <li>
+                     <Translate content="account.votes.total_votes" />: <FormattedAsset amount={total_votes} asset="1.3.0" /><br/>
+                  </li>
+                  <li>
+                     <Translate content="account.votes.daily_pay" />: <FormattedAsset amount={worker.daily_pay} asset="1.3.0" /><br/>
+                  </li>
+                  <li>
+                     <Translate content="account.votes.max_pay" />: <FormattedAsset amount={worker.daily_pay*total_days} asset="1.3.0" /><br/>
+                  </li>
+                  <li>
+                     <Translate content="account.votes.unclaimed" />: <VestingBalance balance={worker.worker[1].balance} /> <br/>
+                  </li>
+                  <li>
+                     <Translate content="account.votes.status.title" />: {approval} <br/>
+                  </li>
+                  <li>
+                     Id: {this.props.worker.get("id")} <br/>
+                  </li>
+               </ul>
+
+               <div className="button-group no-margin" style={{paddingTop: "1rem"}}>
+                  <button className="button success" onClick={this.onApprove.bind(this)}>
+                     <Translate content="account.votes.approve_worker"/>
+                  </button>
+
+                  <button className="button info" onClick={this.onReject.bind(this)}>
+                     <Translate content="account.votes.reject_worker"/>
+                  </button>
+               </div>
             </div>
-
-            <button className="button outline" onClick={this.onApprove.bind(this)}>
-               <Translate content="account.votes.approve_worker"/>
-            </button>
-
-            <button className="button outline" onClick={this.onReject.bind(this)}>
-               <Translate content="account.votes.reject_worker"/>
-            </button>
 
          </div>
       </div>
