@@ -127,7 +127,7 @@ class MyMarkets extends React.Component {
                         <thead>
                             <tr>
                                 <th className="mymarkets-header clickable" onClick={this._inverseSort.bind(this)} style={{textAlign: "left", paddingLeft: "15px"}}><Translate content="exchange.market_name" /></th>
-                                <th className="mymarkets-header" style={{textAlign: "left" }}><button onClick={this._goMarkets.bind(this)} className="button outline"><Translate content="exchange.more" /></button></th>
+                                <th className="mymarkets-header" style={{textAlign: "left" }}><button onClick={this._goMarkets.bind(this)} className="small button outline" style={{padding: "4px 8px", opacity: "0.65" , letterSpacing: "0.05rem"}}><Translate content="exchange.more" /></button></th>
                             </tr>
                         </thead>
                     </table>
