@@ -7,9 +7,11 @@ import ChainTypes from "../Utility/ChainTypes";
 import CachedPropertyStore from "stores/CachedPropertyStore"
 import CachedPropertyActions from "actions/CachedPropertyActions"
 import BlockchainStore from "stores/BlockchainStore";
+import ChainStore from "api/ChainStore"
 import WalletDb from "stores/WalletDb";
 import TimeAgo from "../Utility/TimeAgo";
 import Icon from "../Icon/Icon";
+import ReactTooltip from "react-tooltip"
 
 @BindToChainState({keep_updating: true})
 class Footer extends React.Component {
@@ -22,7 +24,7 @@ class Footer extends React.Component {
     static defaultProps = {
         dynGlobalObject: "2.1.0"
     }
-    
+
     static contextTypes = {
         router: React.PropTypes.func.isRequired
     }
@@ -39,40 +41,30 @@ class Footer extends React.Component {
                nextProps.synced !== this.props.synced;
     }
 
-    _triggerPerf() {
-        if (!this.state.perf) {
-            Perf.start();
-        } else {
-            Perf.stop();
-            console.log("Inclusive prints the overall time taken. If no argument's passed, defaults to all the measurements from the last recording. This prints a nicely formatted table in the console, like so:");
-            Perf.printInclusive();
-            console.log("Wasted time is spent on components that didn't actually render anything, e.g. the render stayed the same, so the DOM wasn't touched.");
-            Perf.printWasted();
-            console.log("Exclusive times don't include the times taken to mount the components: processing props, getInitialState, call componentWillMount and componentDidMount, etc.");
-            Perf.printExclusive();
-            Perf.printDOM();
-        }
-        this.setState({perf: !this.state.perf});
-    }
-
     render() {
         let block_height = this.props.dynGlobalObject.get("head_block_number");
         let block_time = this.props.dynGlobalObject.get("time") + "+00:00";
-        let bt = new Date(block_time).getTime() / 1000;
-        let now = new Date().getTime() / 1000
+        // console.log("block_time", block_time)
+        let bt = (new Date(block_time).getTime() + ChainStore.getEstimatedChainTimeOffset()) / 1000;
+        let now = new Date().getTime() / 1000;
+        let version_match = APP_VERSION.match(/2\.0\.(\d\w+)/);
+        let version = version_match ? `.${version_match[1]}` : ` ${APP_VERSION}`;
         return (
             <div className="show-for-medium grid-block shrink footer">
                 <div className="align-justify grid-block">
-                    <div onClick={this._triggerPerf.bind(this)} className="grid-block">
+                    <div className="grid-block">
                         <div className="logo">
-                            <Translate content="footer.title" />
+                            <Translate content="footer.title" /><span className="version">{version}</span>
                         </div>
                     </div>
                     {this.props.synced ? null : <div className="grid-block shrink txtlabel error">Blockchain is out of sync, please wait until it's synchronized.. &nbsp; &nbsp;</div>}
                     {this.props.rpc_connection_status === "closed" ? <div className="grid-block shrink txtlabel error">No Blockchain connection &nbsp; &nbsp;</div> : null}
                     { this.props.backup_recommended ? <span>
                         <div className="grid-block">
-                            <a className="shrink txtlabel facolor-alert" onClick={this.onBackup.bind(this)}>Backup Required</a>
+                            <a className="shrink txtlabel facolor-alert"
+                                data-tip="Please understand that you are responsible for making your own backup&hellip;"
+                                data-type="warning"
+                                onClick={this.onBackup.bind(this)}>Backup Required</a>
                             &nbsp;&nbsp;
                         </div>
                     </span> : null}
@@ -93,18 +85,18 @@ class Footer extends React.Component {
             </div>
         );
     }
-    
+
     onBackup() {
         this.context.router.transitionTo("wmc-backup-create")
     }
-    
+
     onBackupBrainkey() {
         this.context.router.transitionTo("wmc-backup-brainkey")
     }
 }
 
 class AltFooter extends Component {
-    
+
     render() {
         var wallet = WalletDb.getWallet()
         return <AltContainer
