@@ -13,6 +13,7 @@ import WalletDb from "stores/WalletDb";
 import WalletUnlockStore from "stores/WalletUnlockStore";
 import WalletUnlockActions from "actions/WalletUnlockActions";
 import WalletManagerStore from "stores/WalletManagerStore";
+import cnames from "classnames";
 
 @connectToStores
 class Header extends React.Component {
@@ -30,14 +31,33 @@ class Header extends React.Component {
         }
     }
 
-    shouldComponentUpdate(nextProps) {
+    static contextTypes = {
+        router: React.PropTypes.func.isRequired
+    };
+
+    constructor(props) {
+        super();
+        this.state = {
+            active: null
+        };
+    }
+
+    componentWillMount() {
+        let path = this.context.router.getCurrentPath().replace("/", "");
+        path = path.indexOf("market") !== -1 ? "exchange" : path;
+        this.setState({
+            active: path
+        });
+    }
+
+    shouldComponentUpdate(nextProps, nextState) {
         return (
             nextProps.linkedAccounts !== this.props.linkedAccounts ||
             nextProps.locked !== this.props.locked ||
             nextProps.current_wallet !== this.props.current_wallet ||
-            nextProps.lastMarket !== this.props.lastMarket
+            nextProps.lastMarket !== this.props.lastMarket ||
+            nextState.active !== this.state.active
         );
-
     }
 
     _triggerMenu(e) {
@@ -51,14 +71,32 @@ class Header extends React.Component {
         else WalletUnlockActions.lock();
     }
 
+    _onNavigate(route, e) {
+        e.preventDefault();
+        let path = route.route ? route.route : route;
+        this.setState({active: route.route ? route.route : route});
+        
+        switch(path) {
+            case "exchange":
+                this.context.router.transitionTo(path, route.params);
+                break;
+
+            default:
+                this.context.router.transitionTo(path);
+                break;
+        }
+
+    }
+
     render() {
-        //if(!this.props.current_wallet) return null;
+        let {active} = this.state
         let linkedAccounts = this.props.linkedAccounts;
         let settings = counterpart.translate("header.settings");
         let locked_tip = counterpart.translate("header.locked_tip");
         let unlocked_tip = counterpart.translate("header.unlocked_tip");
         let linkToAccountOrDashboard;
-        if (linkedAccounts.size > 1) linkToAccountOrDashboard = <Link to="dashboard"><Translate component="span" content="header.dashboard" /></Link>;
+
+        if (linkedAccounts.size > 1) linkToAccountOrDashboard = <a className={cnames({active: active === "dashboard"})} onClick={this._onNavigate.bind(this, "dashboard")}><Translate component="span" content="header.dashboard" /></a>;
         else if (linkedAccounts.size === 1) linkToAccountOrDashboard = <Link to="account-overview" params={{account_name: linkedAccounts.first()}}><Translate component="span" content="header.account" /></Link>;
         else linkToAccountOrDashboard = <Link to="create-account">Create Account</Link>;
         let lock_unlock = null;
@@ -69,9 +107,9 @@ class Header extends React.Component {
                 : <a href onClick={this._toggleLock.bind(this)} data-tip={unlocked_tip} data-place="bottom" data-type="light"><Icon name="unlocked"/></a> }
             </div>);
 
-        let tradeLink = this.props.lastMarket ?
-            <Link to="exchange" params={{marketID: this.props.lastMarket}}><Translate component="span" content="header.exchange" /></Link>:
-            <Link to="markets"><Translate component="span" content="header.exchange" /></Link>
+        let tradeLink = this.props.lastMarket && active !== "exchange" ?
+            <a className={cnames({active: active === "exchange" || active === "markets"})} onClick={this._onNavigate.bind(this, {route: "exchange", params: {marketID: this.props.lastMarket}})}><Translate component="span" content="header.exchange" /></a>:
+            <a className={cnames({active: active === "markets" || active === "exchange"})} onClick={this._onNavigate.bind(this, "markets")}><Translate component="span" content="header.exchange" /></a>
         return (
             <div className="header menu-group primary">
                 <div className="show-for-small-only">
@@ -83,10 +121,10 @@ class Header extends React.Component {
                 <div className="show-for-medium medium-8">
                     <ul className="menu-bar">
                         <li>{linkToAccountOrDashboard}</li>
-                        <li><Link to="explorer"><Translate component="span" content="header.explorer" /></Link></li>
+                        <li><a className={cnames({active: active === "explorer"})} onClick={this._onNavigate.bind(this, "explorer")}><Translate component="span" content="header.explorer" /></a></li>
                         {linkedAccounts.size === 0 ? null :
                             <li>{tradeLink}</li>}
-                        <li><Link to="transfer"><Translate component="span" content="header.payments" /></Link></li>
+                        <li><a className={cnames({active: active === "transfer"})} onClick={this._onNavigate.bind(this, "transfer")}><Translate component="span" content="header.payments" /></a></li>
                     </ul>
                 </div>
                 <div className="show-for-medium medium-4">
